+++ controller/cmd/ca/main.go
@@ -16,7 +16,9 @@ import (
 func main() {
 	metricsAddr := flag.String("metrics-addr", ":9997", "address to serve scrapable metrics on")
 	controllerNamespace := flag.String("controller-namespace", "linkerd", "namespace in which Linkerd is installed")
+	singleNamespace := flag.Bool("single-namespace", false, "only operate in the controller namespace")
 	kubeConfigPath := flag.String("kubeconfig", "", "path to kube config")
+	proxyAutoInject := flag.Bool("proxy-auto-inject", false, "if true, watch for the add and update events of mutating webhook configurations")
 	flags.ConfigureAndParse()
 
 	stop := make(chan os.Signal, 1)
@@ -26,14 +28,20 @@ func main() {
 	if err != nil {
 		log.Fatal(err.Error())
 	}
-	k8sAPI := k8s.NewAPI(
-		k8sClient,
-		spClient,
-		k8s.Pod,
-		k8s.RS,
-	)
 
-	controller, err := ca.NewCertificateController(*controllerNamespace, k8sAPI)
+	restrictToNamespace := ""
+	if *singleNamespace {
+		restrictToNamespace = *controllerNamespace
+	}
+
+	var k8sAPI *k8s.API
+	if *proxyAutoInject {
+		k8sAPI = k8s.NewAPI(k8sClient, spClient, restrictToNamespace, k8s.Pod, k8s.RS, k8s.MWC)
+	} else {
+		k8sAPI = k8s.NewAPI(k8sClient, spClient, restrictToNamespace, k8s.Pod, k8s.RS)
+	}
+
+	controller, err := ca.NewCertificateController(*controllerNamespace, k8sAPI, *proxyAutoInject)
 	if err != nil {
 		log.Fatalf("Failed to create CertificateController: %v", err)
 	}