+++ CHANGES.md
@@ -1,3 +1,19 @@
+## edge-18.10.2
+
+This release brings major improvements to the CLI as described below, including
+support for auto-injecting deployments via a Kubernetes Admission Controller.
+Proxy auto-injection is **experimental**, and the implementation may change
+going forward.
+
+* CLI
+  * **New** Added a `--proxy-auto-inject` flag to the `install` command,
+    allowing for auto-injection of sidecar containers (Thanks @ihcsim!)
+  * **Improved** Added `--proxy-cpu` and `--proxy-memory` flags to the `install`
+    and `inject` commands, giving the ability to configure CPU + Memory requests
+    (Thanks @benjdlambert!)
+  * **Improved** Added a `--context` flag to specify the context to use to talk
+    to the Kubernetes apiserver (Thanks @ffd2subroutine!)
+
 ## edge-18.10.1
 
 * Web UI